saved progress

+++ app/Http/Controllers/EventController.php
@@ -2,23 +2,43 @@
 
 namespace App\Http\Controllers;
 
+use Exception;
 use App\Models\Event;
+use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Log;
 use App\Http\Requests\StoreEventRequest;
 use App\Http\Requests\UpdateEventRequest;
-use Exception;
 
 class EventController extends Controller
 {
-    public function index()
+
+    /**
+     * Get paginated events.
+     *
+     * @param \Illuminate\Http\Request $request
+     *
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function index(Request $request)
     {
         try {
-            $events = Event::all();
+            $payload = $request->all();
+            $per_page =  $payload['per_page'] ?? 5;
+            Log::info($payload);
+            $events = Event::paginate($per_page);
             return $this->successResponse($events, 'Events fetched successfully', 200);
         } catch (Exception $e) {
             return $this->errorResponse($e);
         }
     }
 
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \App\Http\Requests\StoreEventRequest  $request
+     * @return \Illuminate\Http\Response
+     */
     public function store(StoreEventRequest $request)
     {
         try {
@@ -29,6 +49,12 @@ class EventController extends Controller
         }
     }
 
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Event  $event
+     * @return \Illuminate\Http\Response
+     */
     public function show(Event $event)
     {
         try {
@@ -38,20 +64,43 @@ class EventController extends Controller
         }
     }
 
-    public function update(UpdateEventRequest $request, Event $event)
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \App\Http\Requests\UpdateEventRequest  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function update(UpdateEventRequest $request)
     {
         try {
+            $id = $request->route('event');
+            $event = Event::findOrFail($id);
             $event->update($request->validated());
+
             return $this->successResponse($event, 'Event updated successfully', 200);
         } catch (Exception $e) {
             return $this->errorResponse($e);
         }
     }
 
-    public function destroy(Event $event)
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy(Request $request)
     {
         try {
+            $id = $request->route('event');
+            $event = Event::find($id);
+            if (!$event) {
+                return $this->errorResponse(null, 'Event not found', 404);
+            }
+
             $event->delete();
+
             return $this->successResponse(null, 'Event deleted successfully', 200);
         } catch (Exception $e) {
             return $this->errorResponse($e);
